+++ packages/lib-react-components/.storybook/main.js
@@ -1,17 +1,7 @@
-const webpack = require('webpack')
-
-const webpackConfig = require('../webpack.dist')
-
 function webpackFinal(config, options) {
-  config.plugins.concat(webpackConfig.plugins)
-
   const resolve = {
     ...config.resolve,
-    alias: {
-      ...webpackConfig.resolve.alias,
-    },
     fallback: {
-      ...webpackConfig.resolve.fallback,
       crypto: false,
       path: 'path-browserify'
     }