+++ packages/lib-classifier/src/components/Classifier/components/FieldGuide/components/FieldGuideButton/HelpIcon.spec.js
@@ -1,9 +1,0 @@
-import { shallow } from 'enzyme'
-import HelpIcon from './HelpIcon'
-
-describe('Component > HelpIcon', function () {
-  it('should render without crashing', function () {
-    const wrapper = shallow(<HelpIcon />)
-    expect(wrapper).to.be.ok()
-  })
-})